+++ packages/stacked_cli/example/lib/ui/views/home/home_viewmodel.dart
@@ -1,8 +1,8 @@
-import 'package:stacked/stacked.dart';
 import 'package:example/app/app.locator.dart';
 import 'package:example/enums/bottom_sheet_type.dart';
 import 'package:example/enums/dialog_type.dart';
 import 'package:example/ui/common/app_strings.dart';
+import 'package:stacked/stacked.dart';
 import 'package:stacked_services/stacked_services.dart';
 
 class HomeViewModel extends BaseViewModel {
@@ -15,7 +15,7 @@ class HomeViewModel extends BaseViewModel {
 
   void incrementCounter() {
     _counter++;
-    notifyListeners();
+    rebuildUi();
   }
 
   void showDialog() {